+++ sample/src/main/kotlin/util/WatchableListSerializer.kt
@@ -23,7 +23,7 @@ class WatchableListSerializer<T : Any>(private val dataSerializer: KSerializer<T
 
     override fun serialize(encoder: Encoder, obj: WatchableList<T>) {
         with(encoder.beginStructure(descriptor)) {
-            runBlocking { obj.get() }.forEachIndexed { index, item ->
+            runBlocking { obj.value }.forEachIndexed { index, item ->
                 encodeSerializableElement(descriptor, index, dataSerializer, item)
             }
             endStructure(descriptor)