+++ src/main/kotlin/io/gladed/watchable/ReadOnlyWatchableSet.kt
@@ -17,4 +17,4 @@
 package io.gladed.watchable
 
 /** A [Watchable] [Set] which may not be modified by the reference holder. */
-interface ReadOnlyWatchableSet<T> : Watchable<Set<T>, SetChange<T>>
+interface ReadOnlyWatchableSet<T> : Watchable<Set<T>, SetChange<T>>, Set<T>