+++ src/main/kotlin/io/gladed/watchable/WatchableList.kt
@@ -19,9 +19,25 @@ package io.gladed.watchable
 /**
  * A [Watchable] wrapper for a [List] which may also be modified or bound. Use [watchableListOf] to create.
  */
+@Suppress("TooManyFunctions")
 class WatchableList<T> internal constructor(
     initial: Collection<T>
 ) : MutableWatchableBase<List<T>, MutableList<T>, ListChange<T>>(), ReadOnlyWatchableList<T> {
+
+    override val size: Int get() = value.size
+    override fun contains(element: T) = value.contains(element)
+    override fun containsAll(elements: Collection<T>) = value.containsAll(elements)
+    override fun get(index: Int) = value[index]
+    override fun indexOf(element: T) = value.indexOf(element)
+    override fun isEmpty() = value.isEmpty()
+    override fun iterator() = value.iterator()
+    override fun lastIndexOf(element: T) = value.lastIndexOf(element)
+    override fun listIterator() = value.listIterator()
+    override fun listIterator(index: Int) = value.listIterator(index)
+    override fun subList(fromIndex: Int, toIndex: Int) = value.subList(fromIndex, toIndex)
+    override fun equals(other: Any?): Boolean = value == other
+    override fun hashCode() = value.hashCode()
+    override fun toString() = "WatchableList($value)"
 
     override val mutable = object : AbstractMutableList<T>() {
         val real = initial.toMutableList()
@@ -74,8 +90,6 @@ class WatchableList<T> internal constructor(
 
     /** Return an unmodifiable form of this [WatchableList]. */
     fun readOnly(): ReadOnlyWatchableList<T> = object : ReadOnlyWatchableList<T> by this {
-        override fun toString() = "ReadOnlyWatchableList()"
+        override fun toString() = "ReadOnlyWatchableList($value)"
     }
-
-    override fun toString() = "WatchableList()"
 }