+++ src/main/kotlin/io/gladed/watchable/WatchableMap.kt
@@ -17,9 +17,22 @@
 package io.gladed.watchable
 
 /** A [Watchable] wrapper for a [Map] which may also be modified or bound. Use [watchableMapOf] to create.*/
+@Suppress("TooManyFunctions")
 class WatchableMap<K, V> internal constructor(
     initial: Map<K, V>
 ) : MutableWatchableBase<Map<K, V>, MutableMap<K, V>, MapChange<K, V>>(), ReadOnlyWatchableMap<K, V> {
+
+    override val entries get() = value.entries
+    override val keys get() = value.keys
+    override val size get() = value.size
+    override val values get() = value.values
+    override fun containsKey(key: K) = value.containsKey(key)
+    override fun containsValue(value: V) = this.value.containsValue(value)
+    override fun get(key: K): V? = value[key]
+    override fun isEmpty() = value.isEmpty()
+    override fun equals(other: Any?) = value == other
+    override fun hashCode() = value.hashCode()
+    override fun toString() = "WatchableMap($value)"
 
     /** A map that checks and reports all change attempts. */
     override val mutable = object : AbstractMutableMap<K, V>() {
@@ -102,8 +115,6 @@ class WatchableMap<K, V> internal constructor(
 
     /** Return an unmodifiable form of this [WatchableMap]. */
     fun readOnly(): ReadOnlyWatchableMap<K, V> = object : ReadOnlyWatchableMap<K, V> by this {
-        override fun toString() = "ReadOnlyWatchableMap()"
+        override fun toString() = "ReadOnlyWatchableMap($value)"
     }
-
-    override fun toString() = "WatchableMap()"
 }